Backport and optimize deadman switch
